+++ test-src/Data/TimeSeries/IO/CSVSpec.hs
@@ -1,5 +1,6 @@
 module Data.TimeSeries.IO.CSVSpec (spec) where
 
+import Prelude
 import Test.Hspec
 import Data.Text.Time (parseISODateTime)